+++ packages/es-dev-server/src/utils/resolve-module-imports.js
@@ -1,0 +1,206 @@
+/* eslint-disable no-await-in-loop, no-restricted-syntax */
+import whatwgUrl from 'whatwg-url';
+import nodeResolve from 'resolve';
+import pathIsInside from 'path-is-inside';
+// typescript can't resolve a .cjs file
+// @ts-ignore
+import { parse } from 'es-module-lexer';
+import path from 'path';
+import { toBrowserPath } from './utils.js';
+
+const CONCAT_NO_PACKAGE_ERROR =
+  'Dynamic import with a concatenated string should start with a valid full package name.';
+
+export class ResolveSyntaxError extends Error {}
+
+/**
+ * @param {string} importPath
+ * @param {import('resolve').AsyncOpts} options
+ * @returns {Promise<string>}
+ */
+function nodeResolveAsync(importPath, options) {
+  return new Promise((resolve, reject) =>
+    nodeResolve(importPath, options, (error, result) => {
+      if (error) {
+        reject(error);
+      } else {
+        resolve(result);
+      }
+    }),
+  );
+}
+
+/**
+ * Resolves an import which is a concatenated string (for ex. import('my-package/files/${filename}'))
+ *
+ * Resolving is done by taking the package name and resolving that, then prefixing the resolves package
+ * to the import. This requires the full package name to be present in the string.
+ *
+ * @param {string} sourceFileDir
+ * @param {string} importPath
+ * @param {object} config
+ * @returns {Promise<string>}
+ */
+async function resolveConcatenatedImport(sourceFileDir, importPath, config) {
+  let pathToResolve = importPath;
+  let pathToAppend = '';
+
+  const parts = importPath.split('/');
+  if (importPath.startsWith('@')) {
+    if (parts.length < 2) {
+      throw new Error(CONCAT_NO_PACKAGE_ERROR);
+    }
+
+    pathToResolve = `${parts[0]}/${parts[1]}`;
+    pathToAppend = parts.slice(2, parts.length).join('/');
+  } else {
+    if (parts.length < 1) {
+      throw new Error(CONCAT_NO_PACKAGE_ERROR);
+    }
+
+    [pathToResolve] = parts;
+    pathToAppend = parts.slice(1, parts.length).join('/');
+  }
+
+  const resolvedPackage = await nodeResolveAsync(`${pathToResolve}/package.json`, {
+    basedir: sourceFileDir,
+    extensions: config.fileExtensions,
+    moduleDirectory: config.moduleDirectories,
+    preserveSymlinks: config.preserveSymlinks,
+
+    packageFilter(packageJson) {
+      /* eslint-disable no-param-reassign */
+      packageJson.main = packageJson.module || packageJson['jsnext:main'] || packageJson.main;
+      return packageJson;
+    },
+  });
+
+  const packageDir = resolvedPackage.substring(0, resolvedPackage.length - 'package.json'.length);
+  return `${packageDir}${pathToAppend}`;
+}
+
+/**
+ * @typedef {object} ResolveConfig
+ * @property {string[]} fileExtensions
+ * @property {string[]} moduleDirectories
+ * @property {boolean} preserveSymlinks
+ */
+
+/**
+ * @param {string} rootDir
+ * @param {string} sourceFilePath
+ * @param {string} importPath
+ * @param {ResolveConfig} config
+ * @param {boolean} [concatenated]
+ */
+async function resolveImport(rootDir, sourceFilePath, importPath, config, concatenated) {
+  // don't touch url imports
+  if (whatwgUrl.parseURL(importPath) !== null) {
+    return importPath;
+  }
+
+  // don't touch relative or absolute imports with a file extension
+  if (
+    (importPath.startsWith('.') || importPath.startsWith('/')) &&
+    path.extname(importPath) !== ''
+  ) {
+    return importPath;
+  }
+
+  const sourceFileDir = path.dirname(sourceFilePath);
+  try {
+    let resolvedImportFilePath;
+
+    if (concatenated) {
+      // if this dynamic import is a concatenated string, try our best to resolve. Otherwise leave it untouched and resolve it at runtime.
+      try {
+        resolvedImportFilePath = await resolveConcatenatedImport(sourceFileDir, importPath, config);
+      } catch (error) {
+        return importPath;
+      }
+    } else {
+      resolvedImportFilePath = await nodeResolveAsync(importPath, {
+        basedir: sourceFileDir,
+        extensions: config.fileExtensions,
+        moduleDirectory: config.moduleDirectories,
+        preserveSymlinks: config.preserveSymlinks,
+        packageFilter(packageJson) {
+          /* eslint-disable no-param-reassign */
+          packageJson.main = packageJson.module || packageJson['jsnext:main'] || packageJson.main;
+          return packageJson;
+        },
+      });
+    }
+
+    if (!pathIsInside(resolvedImportFilePath, rootDir)) {
+      throw new Error(
+        `Import "${importPath}" resolved to the file "${resolvedImportFilePath}" which is outside the root directory. ` +
+          'Install the module locally in the current project, or expand the root directory. If this is a symlink or if you used npm link, ' +
+          ' you can use the preserveSymlinks option',
+      );
+    }
+
+    const relativeImportFilePath = path.relative(sourceFileDir, resolvedImportFilePath);
+    const resolvedImportPath = toBrowserPath(relativeImportFilePath);
+    return resolvedImportPath.startsWith('.') ? resolvedImportPath : `./${resolvedImportPath}`;
+  } catch (error) {
+    // make module not found error message shorter
+    if (error.code === 'MODULE_NOT_FOUND') {
+      throw new Error(`Could not find module "${importPath}".`);
+    }
+    throw error;
+  }
+}
+
+/**
+ * @param {string} rootDir
+ * @param {string} sourceFilePath
+ * @param {string} source
+ * @param {ResolveConfig} config
+ */
+export async function resolveModuleImports(rootDir, sourceFilePath, source, config) {
+  let imports;
+  try {
+    [imports] = await parse(source);
+  } catch (error) {
+    throw new ResolveSyntaxError('Syntax error.');
+  }
+
+  let resolvedSource = '';
+  let lastIndex = 0;
+
+  for (const imp of imports) {
+    const { s: start, e: end, d: dynamicImportIndex } = imp;
+
+    if (dynamicImportIndex === -1) {
+      // static import
+      const importPath = source.substring(start, end);
+      const resolvedImportPath = await resolveImport(rootDir, sourceFilePath, importPath, config);
+
+      resolvedSource += `${source.substring(lastIndex, start)}${resolvedImportPath}`;
+      lastIndex = end;
+    } else if (dynamicImportIndex >= 0) {
+      // dynamic import
+      const dynamicStart = start + 1;
+      const dynamicEnd = end - 1;
+
+      const importPath = source.substring(dynamicStart, dynamicEnd);
+      const stringSymbol = source[dynamicStart - 1];
+      const isStringLiteral = [`\``, "'", '"'].includes(stringSymbol);
+      const dynamicString =
+        stringSymbol === `\`` || importPath.includes("'") || importPath.includes('"');
+      const resolvedImportPath = isStringLiteral
+        ? await resolveImport(rootDir, sourceFilePath, importPath, config, dynamicString)
+        : importPath;
+
+      resolvedSource += `${source.substring(lastIndex, dynamicStart)}${resolvedImportPath}`;
+      lastIndex = dynamicEnd;
+    }
+  }
+
+  if (lastIndex < source.length - 1) {
+    resolvedSource += `${source.substring(lastIndex, source.length)}`;
+  }
+
+  return resolvedSource;
+}