+++ packages/es-dev-server/test/snapshots/resolve-module-imports/import-in-code.js
@@ -1,0 +1,18 @@
+
+      function myimport() {
+
+      }
+
+      function my_import() {
+
+      }
+
+      function importShim() {
+
+      }
+
+      class Foo {
+        import() {
+          return 'foo';
+        }
+      }