+++ packages/es-dev-server/test/snapshots/resolve-module-imports/string-concat-errors.js
@@ -1,0 +1,6 @@
+const file = 'a';
+       import(`@namespace/non-existing/dynamic-files/${file}.js`);
+       import(`non-existing/dynamic-files/${file}.js`);
+       import(totallyDynamic);
+       import(`${file}.js`);
+