+++ packages/es-dev-server/test/snapshots/resolve-module-imports/string-concatenation.js
@@ -1,0 +1,7 @@
+const file = 'a';
+       import(`../node_modules/@namespace/my-module-3/dynamic-files/${file}.js`);
+       import(`../node_modules/my-module/dynamic-files/${file}.js`);
+       import('../node_modules/my-module/dynamic-files' + '/' + file + '.js');
+       import("../node_modules/my-module/dynamic-files/" + file + ".js");
+       import('../node_modules/my-module/dynamic-files'.concat(file).concat('.js'));
+