+++ packages/karma-esm/demo/typescript/karma.config.js
@@ -48,7 +48,7 @@ module.exports = config => {
       babelModernExclude: [
         '**/node_modules/sinon/**/*',
         '**/node_modules/mocha/**/*',
-        '**/node_modules/@bundled-es-modules/chai/**/*',
+        '**/node_modules/chai/**/*',
         '**/node_modules/sinon/chai/**/*',
       ],
       fileExtensions: ['.ts'],