+++ packages/semantic-dom-diff/chai-dom-diff.d.ts
@@ -1,0 +1,7 @@
+/// <reference path="chai-dom-diff-plugin.d.ts" />
+
+/**
+ * @param {any} chai
+ * @param {any} utils
+ */
+export const chaiDomDiff: (chai: any, utils: any) => void;