+++ packages/semantic-dom-diff/index.d.ts
@@ -1,0 +1,3 @@
+export { chaiDomDiff } from "./chai-dom-diff.js";
+import { getDiffableHTML } from "./get-diffable-html.js";
+export { getDiffableHTML, getDiffableHTML as getDiffableSemanticHTML };