+++ packages/semantic-dom-diff/test/bdd-setup.js
@@ -1,6 +1,6 @@
 /* istanbul ignore next */
 // do manual setup and not use testing to not have circle dependencies
-import { chai } from '@bundled-es-modules/chai';
+import 'chai/chai.js';
 import { cachedWrappers } from '@open-wc/testing-helpers/index-no-side-effects.js';
 import { chaiDomDiff } from '../chai-dom-diff.js';
 
@@ -17,4 +17,9 @@ if (afterEach) {
 }
 
 // register-plugins
-chai.use(chaiDomDiff);
+// @ts-ignore
+window.chai.use(chaiDomDiff);
+
+// @ts-ignore
+const { expect, assert, should } = window.chai;
+export { expect, assert, should };