+++ packages/testing/register-chai-plugins.js
@@ -1,6 +1,12 @@
-import { chai } from '@bundled-es-modules/chai';
+// non es module plugins
+import 'chai-dom';
+import 'sinon-chai';
+
+// es module plugins
 import { chaiDomDiff } from '@open-wc/semantic-dom-diff';
 import { chaiA11yAxe } from 'chai-a11y-axe';
 
-chai.use(chaiDomDiff);
-chai.use(chaiA11yAxe);
+// @ts-ignore
+window.chai.use(chaiDomDiff);
+// @ts-ignore
+window.chai.use(chaiA11yAxe);