+++ packages/building-utils/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.1](https://github.com/open-wc/open-wc/compare/@open-wc/building-utils@2.8.0...@open-wc/building-utils@2.8.1) (2019-08-21)
+
+
+### Bug Fixes
+
+* **building-utils:** re-add removed file ([1872515](https://github.com/open-wc/open-wc/commit/1872515))
+
+
+
+
+
+# [2.8.0](https://github.com/open-wc/open-wc/compare/@open-wc/building-utils@2.7.0...@open-wc/building-utils@2.8.0) (2019-08-20)
+
+
+### Features
+
+* **es-dev-server:** faster node-resolve via use of es-module-lexer ([9589e4e](https://github.com/open-wc/open-wc/commit/9589e4e))
+
+
+
+
+
 # [2.7.0](https://github.com/open-wc/open-wc/compare/@open-wc/building-utils@2.6.1...@open-wc/building-utils@2.7.0) (2019-08-14)