+++ packages/testing-karma/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.24](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma@3.1.23...@open-wc/testing-karma@3.1.24) (2019-08-22)
+
+**Note:** Version bump only for package @open-wc/testing-karma
+
+
+
+
+
+## [3.1.23](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma@3.1.22...@open-wc/testing-karma@3.1.23) (2019-08-21)
+
+**Note:** Version bump only for package @open-wc/testing-karma
+
+
+
+
+
+## [3.1.22](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma@3.1.21...@open-wc/testing-karma@3.1.22) (2019-08-20)
+
+**Note:** Version bump only for package @open-wc/testing-karma
+
+
+
+
+
+## [3.1.21](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma@3.1.20...@open-wc/testing-karma@3.1.21) (2019-08-20)
+
+
+### Bug Fixes
+
+* do not destructure exports to support es-module-lexer ([3709413](https://github.com/open-wc/open-wc/commit/3709413))
+
+
+
+
+
+## [3.1.20](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma@3.1.19...@open-wc/testing-karma@3.1.20) (2019-08-18)
+
+
+### Bug Fixes
+
+* use chai instead of @bundled-es-modules/chai ([f9d19bb](https://github.com/open-wc/open-wc/commit/f9d19bb))
+
+
+
+
+
 ## [3.1.19](https://github.com/open-wc/open-wc/compare/@open-wc/testing-karma@3.1.18...@open-wc/testing-karma@3.1.19) (2019-08-17)
 
 **Note:** Version bump only for package @open-wc/testing-karma