+++ packages/testing/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.2](https://github.com/open-wc/open-wc/compare/@open-wc/testing@2.3.1...@open-wc/testing@2.3.2) (2019-08-20)
+
+
+### Bug Fixes
+
+* do not destructure exports to support es-module-lexer ([3709413](https://github.com/open-wc/open-wc/commit/3709413))
+
+
+
+
+
+## [2.3.1](https://github.com/open-wc/open-wc/compare/@open-wc/testing@2.3.0...@open-wc/testing@2.3.1) (2019-08-18)
+
+
+### Bug Fixes
+
+* include *.ts files in npm packages ([8087906](https://github.com/open-wc/open-wc/commit/8087906))
+
+
+
+
+
+# [2.3.0](https://github.com/open-wc/open-wc/compare/@open-wc/testing@2.2.8...@open-wc/testing@2.3.0) (2019-08-18)
+
+
+### Features
+
+* **testing:** use chai instead of @bundled-es-modules/chai ([53579c2](https://github.com/open-wc/open-wc/commit/53579c2))
+* add type definition files for testing ([462a29f](https://github.com/open-wc/open-wc/commit/462a29f))
+
+
+
+
+
 ## [2.2.8](https://github.com/open-wc/open-wc/compare/@open-wc/testing@2.2.7...@open-wc/testing@2.2.8) (2019-08-07)
 
 **Note:** Version bump only for package @open-wc/testing